move service-discover to internal/pkg
