+++ internal/pkg/service-discover/erda-discover/provider.go
@@ -19,7 +19,8 @@ import (
 	"net/url"
 
 	"github.com/erda-project/erda-infra/base/servicehub"
-	servicediscover "github.com/erda-project/erda/providers/service-discover"
+	servicediscover "github.com/erda-project/erda/internal/pkg/service-discover"
+	"github.com/erda-project/erda/pkg/discover"
 )
 
 type config struct {
@@ -54,7 +55,7 @@ func (p *provider) Endpoint(scheme, service string) (string, error) {
 			return u.Host, nil
 		}
 	}
-	return "", fmt.Errorf("not found endpoint %q", service)
+	return discover.GetEndpoint(service)
 }
 
 func (p *provider) ServiceURL(scheme, service string) (string, error) {
@@ -63,14 +64,20 @@ func (p *provider) ServiceURL(scheme, service string) (string, error) {
 			return u.String(), nil
 		}
 	}
-	return "", fmt.Errorf("not found %s url for service %q", scheme, service)
+	addr, err := discover.GetEndpoint(service)
+	if err != nil {
+		return "", err
+	}
+	return fmt.Sprintf("%s://%s", scheme, addr), nil
 }
 
 func init() {
-	servicehub.Register("fixed-discover", &servicehub.Spec{
+	servicehub.Register("erda-discover", &servicehub.Spec{
 		Services:    []string{"discover"},
 		Description: "discover all services",
 		ConfigFunc:  func() interface{} { return &config{} },
-		Creator:     func() servicehub.Provider { return &provider{} },
+		Creator: func() servicehub.Provider {
+			return &provider{}
+		},
 	})
 }
