+++ internal/pkg/service-discover/interface.go
