+++ internal/tools/openapi/openapi-ng/example/publish/provider.go
@@ -18,9 +18,9 @@ import (
 	"github.com/erda-project/erda-infra/base/logs"
 	"github.com/erda-project/erda-infra/base/servicehub"
 	transhttp "github.com/erda-project/erda-infra/pkg/transport/http"
+	discover "github.com/erda-project/erda/internal/pkg/service-discover"
 	"github.com/erda-project/erda/internal/tools/openapi/openapi-ng"
 	"github.com/erda-project/erda/internal/tools/openapi/openapi-ng/proxy"
-	discover "github.com/erda-project/erda/providers/service-discover"
 )
 
 // +provider
